+++ packages/mdx/package.json
@@ -48,7 +48,7 @@
     "@babel/core": "^7.0.0",
     "@babel/plugin-transform-react-jsx": "^7.1.6",
     "@mapbox/rehype-prism": "^0.3.0",
-    "@mdx-js/tag": "^1.0.0-alpha.6",
+    "@mdx-js/react": "^1.0.0-alpha.6",
     "hast-util-select": "^3.0.0",
     "jest": "^23.6.0",
     "prettier": "^1.14.2",
