+++ packages/mdx/test/index.test.js
@@ -6,8 +6,8 @@ const prism = require('@mapbox/rehype-prism')
 const math = require('remark-math')
 const katex = require('rehype-katex')
 const prettier = require('prettier')
-const {MDXProvider} = require('@mdx-js/tag')
-const createElement = require('../create-element')
+const {MDXProvider} = require('@mdx-js/react')
+const createElement = require('@mdx-js/react/create-element')
 const React = require('react')
 const {renderToStaticMarkup} = require('react-dom/server')
