+++ packages/react/create-element.js
@@ -1,0 +1,3 @@
+const createElement = require('./dist/create-element').default
+
+module.exports = createElement
