+++ packages/react/test/fixture.js
@@ -1,0 +1,11 @@
+/* @jsx mdx */
+
+// eslint-disable-next-line no-unused-vars
+import mdx from '../src/create-element'
+
+export default () => (
+  <wrapper>
+    <h1 />
+    <h2 />
+  </wrapper>
+)
