+++ packages/react/test/test.js
@@ -1,7 +1,8 @@
 import React from 'react'
 import {renderToString} from 'react-dom/server'
 
-import {MDXTag, MDXProvider} from '../src'
+import {MDXProvider} from '../src/context'
+import Fixture from './fixture'
 
 const H1 = props => <h1 style={{color: 'tomato'}} {...props} />
 const H2 = props => <h2 style={{color: 'rebeccapurple'}} {...props} />
@@ -9,31 +10,17 @@ const CustomH2 = props => <h2 style={{color: 'papayawhip'}} {...props} />
 
 it('Should allow components to be passed via context', () => {
   const Layout = ({children}) => <div id="layout">{children}</div>
-  const components = {h1: H1}
+
   const result = renderToString(
-    <MDXProvider components={components}>
-      <MDXTag Layout={Layout} name="wrapper">
-        <MDXTag name="h1" />
-      </MDXTag>
+    <MDXProvider components={{h1: H1, wrapper: Layout}}>
+      <Fixture />
     </MDXProvider>
   )
 
   // Layout is rendered
   expect(result).toMatch(/id="layout"/)
 
-  // MDXTag picks up on component context
-  expect(result).toMatch(/style="color:tomato"/)
-})
-
-it('Should allow context components to be overridden', () => {
-  const components = {h1: H1}
-  const result = renderToString(
-    <MDXProvider components={{}}>
-      <MDXTag name="h1" components={components} />
-    </MDXProvider>
-  )
-
-  // MDXTag is passed overriding components
+  // H1 is rendered
   expect(result).toMatch(/style="color:tomato"/)
 })
 
@@ -43,8 +30,7 @@ it('Should merge components when there is nested context', () => {
   const result = renderToString(
     <MDXProvider components={components}>
       <MDXProvider components={{h2: CustomH2}}>
-        <MDXTag name="h1" />
-        <MDXTag name="h2" />
+        <Fixture />
       </MDXProvider>
     </MDXProvider>
   )
@@ -62,8 +48,7 @@ it('Should allow removing of context components using the functional form', () =
   const result = renderToString(
     <MDXProvider components={components}>
       <MDXProvider components={_outerComponents => ({h2: CustomH2})}>
-        <MDXTag name="h1" />
-        <MDXTag name="h2" />
+        <Fixture />
       </MDXProvider>
     </MDXProvider>
   )
