--- packages/tag/src/index.js
@@ -1,7 +1,0 @@
-export {default as MDXTag} from './mdx-tag'
-export {
-  default as MDXContext,
-  MDXProvider,
-  withMDXComponents,
-  useMDXComponents
-} from './mdx-context'
